// src/types/types.ts
export type TableOfContentsProps = {
  activeSection: string
}
export type PaperSerctionProps = {
  id: string;
  number: string;
  title: string;
  children: any;
}
export type PaperInfoCardProps = {
  title: string;
  value: string;
  subtext: string;
  icon: any;
}
export type PaperRoadmapItemProps = {
  title: string;
  date: string;
  desc: string;
  status: any;
}